Improve the performance of some functions in ArrayModule
Operations on basic types should be as fast as possible, so that
programmers won't feel the need to re-implement them manually and will
instead be able to use them as building blocks for more complex
functions, without significant performance losses.

Instead of using optimized closures, trivial operations on arrays can
result in much better performance when defined as inline functions,
because this often opens up further optimization opportunities to the
compiler.

See https://github.com/fsharp/fsharp/issues/176 for the benchmarking
code and a more complete discussion.

WARNING: ONLY FOR TESTING, DO NOT MERGE.

Currently this commit seems to work on top of master (b3e994b) and
fsharp_31 (fec482c), but it results in a weird behavior of the build
process:

 - on master, everything seems to work just fine

 - on fsharp_31 bootstrapping and installing on a clean mono works
   fine; after installing, re-building fails with an internal error

/Users/ranma42/Code/sharp/fsharp/src/fsharp/FSharp.Core/FSharp.Core.fsproj (default targets) ->
/Users/ranma42/Code/sharp/fsharp/lib/bootstrap/4.0/Microsoft.FSharp.Targets (CoreCompile target) ->

: error FS0193: internal error: Could not load type 'ErrorStrings' from assembly 'FSharp.Compiler-proto'.Stack TraceSystem.TypeLoadException: Could not load type 'ErrorStrings' from assembly 'FSharp.Compiler-proto'.  at Microsoft.FSharp.Compiler.Tast+RecdFieldRef.get_Index () [0x00000] in <filename unknown>:0
(x8)